♻️ Move log types into API types file
